requirements.txt

liac-arff==2.5.0
matplotlib-inline==0.1.3
matplotlib==3.5.2
numpy==1.22.3
pandas==1.4.2
scikit-learn==1.0.2
seaborn==0.11.2
